requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pymupdf
python-docx
langchain
transformers
torch
accelerate
bitsandbytes